Adjust css import warning text
I saw this while using dynamic import on a css file that had an `@import` statement.

The previous error message has a grammar error, but also says that the rules are not taking effect. However, in my case the import seemed to work, despite the warning.

The explanatory comment above the warning seemed clearer to me and also doesn't claim that it is definitely an error, so I reused that text in the warning itself.